Add restore note revision action generator
