+++ lib/state/data/reducer.ts
@@ -108,7 +108,7 @@ export const notes: A.Reducer<Map<T.EntityId, T.Note>> = (
 
     case 'NOTE_BUCKET_UPDATE':
     case 'REMOTE_NOTE_UPDATE':
-    case 'APPLY_NOTE_REVISION':
+    case 'RESTORE_NOTE_REVISION':
       return new Map(state).set(action.noteId, action.note);
 
     case 'IMPORT_NOTE_WITH_ID': {
@@ -437,7 +437,7 @@ export const tags: A.Reducer<Map<T.TagHash, T.Tag>> = (
       return next;
     }
 
-    case 'APPLY_NOTE_REVISION': {
+    case 'RESTORE_NOTE_REVISION': {
       const next = new Map(state);
       action.note.tags.forEach((tagName) => {
         const tagHash = t(tagName);
